Imputed and transformed data extract from the SMART study

Data from the SMART study, a prospective cohort study at the University Medical Centre Utrecht, the Netherlands. The study was designed to (a) establish the prevalence of concomitant arterial diseases and risk factors for cardiovascular disease in a high-risk population; (b) identify predictors of future cardiovascular events in patients with symptomatic cardiovascular disease.

Usage

data("SMARTfull")

Format

A data frame with 3873 observations on the following 16 variables.

TEVENT

a numeric vector. Time to cardiovascular event (days)

EVENT

a logical vector. Cardiovascular event (clinical, 0/1)

AGE.TRANS

a numeric vector. (Age(years)-50)^2 if over 50, otherwise 0

SEX

a numeric vector. 1 Male 2 Female

SMOKING

a factor with levels 1 Never 2 Former 3 Current

ALCOHOL

a factor with levels 1 Never 2 Former 3 Current

BMI

a numeric vector (imputed). Body mass index (kg/m^2)

SYSTH

a numeric vector (imputed). Systomlic blood pressure (mm/Hg)

HDL

a numeric vector (imputed). High-density lipoprotein cholesterol (mmol/L)

DIABETES

a numeric vector (imputed). Ever diabetes (0/1)

HISTCAR2

a numeric vector. History of cardiac disease score.

HOMOC

a numeric vector (imputed). Homocysteine (\(\mu\)mol/L

logCREAT

a numeric vector (imputed). Creatinine clearance (log mL/min)

ALBUMIN

a factor with levels 1 No 2 Micro 3 Macro. Albumin in urine

STENOSIS

a numeric vector (imputed). Carotic stenosis \(\ge 50\%\) by duplex

IMT

a numeric vector (imputed). Intima media thickness (mm)

Details

SMARTfull is a processed extract from the SMART study. The code for processing is available: http://www.clinicalpredictionmodels.org/doku.php?id=rcode_and_data:chapter23. This dataset is named SMARTc in that code, and is used to fit the "full" model, which is discussed in Chapter 23 of Steyerberg's Clinical Prediction Models.

Source

http://www.clinicalpredictionmodels.org

References

Steyerberg, E.W. (2009) Clinical Prediction Models. New York: Springer-Verlag

Examples

data(SMARTfull)