Fit Cox models using glmBayesMfp

A simplified formula based interface to glmBayesMfp to fit Cox models. Can return Maximum a posteriori (MAP) model, Median probability model (MPM) or Bayesian model average (BMA). Provides global empirical Bayes and AIC/BIC based model inference.

Usage

coxTBF(
  formula,
  data,
  type = "MAP",
  baseline = "shrunk",
  globalEB = FALSE,
  IC = FALSE,
  sep = FALSE,
  keepModelList = FALSE,
  ...,
  overrideConfig
)

Arguments

formula

model formula with Surv object as LHS and uc or bfp variables as RHS.

data

data.frame for model variables

type

type of model to fit, one of "MAP","MPM","BMA","BMAFull"

baseline

how to calculate the baseline hazard function. "cox" uses unshrunken coefficients. "shrunk" refits baseline with shrunken coefficients (default).

globalEB

use global empirical bayes estimate of g (default=FALSE)

IC

use information criteria based model selection (default=FALSE). Either "AIC" or "BIC".

sep

estimate baseline hazard for each estimate of model coefficients (default=FALSE).

keepModelList

keep the model list returned by glmBayesMfp for MAP and MPM models (default=FALSE).

...

additional arguments to pass to glmBayesMfp

overrideConfig

replaces the the MAP model with the given configuration, which is passed to computeModels

Value

An object of S3 class TBFcox or TBFcox.sep if sep=TRUE.