Test the Cox model computation for the TBF approach

Test the Cox model computation for the TBF approach

Usage

testCox(
  survTimes,
  censInd,
  offsets = rep.int(0, length(survTimes)),
  X,
  useCppCode = FALSE
)

Arguments

survTimes

the survival times

censInd

the logical censoring indicator (TRUE = observed, FALSE = censored survival times)

offsets

known component to be included in the linear predictor during fitting. This must be a numeric vector of length equal to the number of cases (if not provided, a vector of zeroes)

X

the design matrix, *without* the intercept 1's!, with at least one column

useCppCode

use the C++ code? (default) otherwise the R-function coxph is used

Value

a list with the coefficient estimates (betas), the covariance matrix estimate (cov) and the residual deviance (deviance).

Author

Daniel Sabanes Bove